バッテリーの％表示

+++ Calibration.cpp
@@ -109,7 +109,7 @@ void Calibration::drawLCD(){
     sprintf(gra, "gray = %d", gray);
     sprintf(gre, "green = %d", green);
     sprintf(tar, "target = %d", target);
-    sprintf(bat, "battery = %d %%", ev3_battery_voltage_mV() / 9400 * 100);
+    sprintf(bat, "battery = %d %%", (int)((double)ev3_battery_voltage_mV() / 9400.0 * 100.0));
 
     ev3_lcd_draw_string(w, 0, 0);
     ev3_lcd_draw_string(b, 0, 8);
